fix: rename lib files to use underscores for Convex compatibility
Convex module paths cannot contain hyphens, only alphanumeric
characters, underscores, or periods. Renamed:
- ai-client.ts → ai_client.ts
- message-builder.ts → message_builder.ts
- message-service.ts → message_service.ts

Updated imports to use new filenames.
