+++ convex/lib/ai_client.ts
