+++ convex/lib/message_builder.ts
